render: extract shared event renderer package
Moves the v0.5.1 renderer machinery (Renderer struct + methods, EventRow,
EventsPage, FetchEvents, LoadLabels, PadDisplay) out of internal/humancli
into a new internal/render package, taking a small Caller interface instead
of humancli's callData directly so transport plumbing stays in exactly one
place per consumer. humancli's events.go/watch.go now consume internal/render
through thin type aliases and a callDataCaller adapter; behavior is
byte-identical (all existing humancli tests pass unchanged in what they
assert, only renamed Header/Line call sites to match the exported API).
Prepares internal/station (task 7) to reuse the same renderer for its feed
pane without a second copy of the vocabulary.
